refactor: Split out meter logic into useMeter hook
